pyproject.toml
[project]
name = "mqtt-mcp"
version = "0.1.4"
description = "A Model Context Protocol (MCP) server for MQTT"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.5.1",
    "paho-mqtt>=2.1.0",
    "pydantic-settings>=2.9.1",
]

[project.scripts]
mqtt-mcp = "mqtt_mcp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=1.0.0",
    "ruff>=0.11.11",
]
